feat: 투표 카테고리 필터 기능 구현

+++ data/src/main/java/com/susu/data/data/repository/VoteRepositoryImpl.kt
@@ -33,7 +33,7 @@ class VoteRepositoryImpl @Inject constructor(
         content: String?,
         mine: Boolean?,
         sortType: String?,
-        categoryId: Long?,
+        categoryId: Int?,
         page: Int?,
         size: Int?,
         sort: String?,
