+++ domain/src/main/java/com/susu/domain/repository/VoteRepository.kt
@@ -13,7 +13,7 @@ interface VoteRepository {
         content: String?,
         mine: Boolean?,
         sortType: String?,
-        categoryId: Long?,
+        categoryId: Int?,
         page: Int?,
         size: Int?,
         sort: String?,
